requirements.txt
certifi==2022.5.18.1
charset-normalizer==2.0.12
colorzero==2.0
idna==3.3
requests==2.27.1
RPi.GPIO==0.7.1
urllib3==1.26.9